renamed widget to be uniform
